requirements.txt
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
django-cors-headers==2.3.0
django-filter==2.1.0
django-sslserver==0.20
djangorestframework==3.9.1
idna==2.8
Markdown==3.0.1
pkg-resources==0.0.0
psycopg2==2.7.5
python-decouple==3.1
pytz==2018.9
requests==2.21.0
soupsieve==1.7.3
urllib3==1.24.1
